requirements.txt
kafka-python
scikit-learn
pandas
numpy
plotly
dash
pymongo
schedule
